@@ -29,8 +29,8 @@ const PostCard = ({ postData }) => {
         <div className="text-xl">{postData.title}</div>
         <div className="text-gray-400 text-sm">{postData.travel_date}</div>
         <div className="mt-auto ml-auto mb-1.5 py-1 w-28 bg-gray-200 hover:bg-gray-300 duration-200 rounded-md text-center text-sm">
-          <Link href={`/travel/posts/${postData.slug}`} passHref>
-            <a>Read more...</a>
+          <Link href={`/travel/posts/${postData.slug}`}>
+            Read more...
           </Link>
         </div>
       </div>
